added long int handling, but have floating point error on cluster

+++ prefAttachModel.h
@@ -33,7 +33,7 @@ class prefAttachModel {
       void saveData(std::vector< T > &data, std::ofstream &fileHandle);
     void save_coeffs(const std::vector< std::vector< double > > &data, std::ofstream &fileHandle);
  public:
-    void run(long int nSteps, int dataInterval);
+    void run(long int nSteps, long int dataInterval);
     std::ofstream &createFile(const std::string base, const std::string dir, std::vector<double> &addtnlData, std::vector<std::string> &addtnlDataLabels);
     prefAttachModel(int n, int m, double kappa);
     ~prefAttachModel() {
